Create symbols for generics at the start of the scope
Generic interfaces can have the same name as contained subprograms
or derived types. This means a subprogram and derived type can have
the same name if there is a generic interface with that name.
Change the order of processing declarations to do generics first
so that we don't report a conflict between the subprogram and type.
(See `module m4b` in `test/semantics/resolve18.f90` for an example.)

`PreSpecificationConstruct` is called on each `SpecificationConstruct`
of a specification part before we continue walking it. We use that
to call `CreateGeneric` for each generic interface so that they already
exist when we do the normal walk of the declarations.

Improve the error message issued when a procedure and generic interface
have the same name but the procedure is not a specific of the generic.

Change `SayAlreadyDeclared` to report the error on the second occurence
of the name when possible. This can arise for declarations the are
processed out of order, e.g. contained subprograms and generic
interfaces.

Avoid multiple "already declared" errors for the case when a contained
subprogram has the same name as a declared entity. We first create the
symbol with SubprogramNameDetails, then replace it with the entity
(and report the error), then replace it with the real subprogram
(and get the error again). By setting and checking the error flag
we avoid the second error.